requirements.txt
requests
pybaseball
ncaa-bbStats
ncaa-stats-py
pandas
python-dotenv
python-telegram-bot
scipy
streamlit==1.32.2
altair==4.2.2
PyGithub
numpy
st-gsheets-connection
pulp
telegram
apscheduler
psycopg2-binary
extra-streamlit-components==0.1.60
